Use python exec from virtualenv
This is no longer easily testable by unit tests, add an integration test
that can be run to test virtualenvs and externalPylint.

Only now when writing this do I realize that it could have been tested by
unit tests by mocking grunt.util.spawn.. Oh well, that's left as an
excercise to the reader of the git log.